fixes truncated uint64_t value in put_unaligned_uint64_t() function
(issue #17804)


git-svn-id: https://origsvn.digium.com/svn/asterisk/branches/1.8@282543 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ include/asterisk/unaligned.h
@@ -50,7 +50,7 @@ static inline unsigned short get_unaligned_uint16(const void *p)
 
 static inline void put_unaligned_uint64(void *p, uint64_t datum)
 {
-	struct { unsigned int d; } __attribute__((packed,may_alias)) *pp = p;
+	struct { uint64_t d; } __attribute__((packed,may_alias)) *pp = p;
 
 	pp->d = datum;
 }
